feat: rewrite git management with sqlite
use `tami5/sqlite.nvim` to store git informations. Run git small
commands in system and run git status management in a libuv job. This
allows to fetch millions of status and add them partially in the
database to avoid destroying the lua memory with huge tables. The paths
can then be retrieved partially to apply updates.